Extract `normalize` into its own submodule
